scripts:pyesw:Add support to select the library from custom repo
Add support to include libraries from a custom repository.
If a user wishes to add a library from a specified path, they can
do so by providing the library name and path in the format "library:library_path",
where the library name is separated from the path by a colon(:).
This feature will not affect the current process, if the user only provides
the library name to the "-al" option, the library will be added from the default repository.
Sample syntax:
config_bsp -al library libmetal:<path to library source>

Signed-off-by: Sathish Kumar Kamishettigari <sathishkumar.kamishettigari@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>