xilpm_ng: runtime: Clear power down status for the core and simplify ack processing
Clear out the power down status for APU cores during core power
down, we want to reset the state of the core during shutdown and it
equivalent to initial boot.
This is essential for the subsystem restart to work correctly and all
upgrade scenarios to work correctly.

Also simplify the ACK processing logic in PM_FORCE_POWERDOWN call.

Signed-off-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>

Acked-for-series:Trung Tran<trung.tran@amd.com>