translation_table: versal_2ve_2vm: Fix incorrect loop count in MMUTableL1
The loop count was incorrectly set to 2 instead of 28 at address 0x1_0000_0000
and resulting in translation fault as the SECT was set to incorrect value
of 0x1_8000_0000. The value of SECT was expected to be 0x8_0000_0000 before starting
traslation table entries for DDR region XPAR_AXI_NOC2_DDR_LOW_1_BASEADDR which
starts at 0x8_0000_0000.

Also, reset the UNDEF_START to 0 before using it for any DDR region to
prevent possible corruption of translation table in certain scenarios.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Anirudha Sarangi <anirudha.sarangi@amd.com>