xilloader: Remove checksum validation for versal_2ve_2vm
For versal and versalnet, checksum should not be enabled
along with authentication and encryption. But for versal_2ve_2vm,
due to new PDI format, checksum needs to be allowed always even when
authetication and encryption are enabled.
Bypassing the check for versal_2ve_2vm in PLM to allow the same.

Signed-off-by: Kalyani Akula <kalyani.akula@amd.com>
Acked-by: Durga Challa <durga.challa@amd.com>