lib: sw_apps: memory_tests: Add warning message about holes in DDR
The existing memory test app doesnt take care of holes in the given DDR
region and generates memory test config with lowest of all base addresses
as start address and highest of all high address as end address and this is
resulting in memory test hang for regions with holes. Hence, add the warning
message about this behavior in the memory tests app.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Anirudha Sarangi <anirudha.sarangi@amd.com>