xilpm_ng: xilpm_boot: fix APU cluster power down
Currently check for APU core power state to decide power down cluster is
incorrect. Because of that APU cluster is not powering down when all cores
in the cluster are powered off. Fix this check by correcting bits for core
state to decide power down of cluster.

Signed-off-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>

Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>