Revert "scripts:pyesw:library_utils:Add support to read the examples from the lib_list.yaml"
This reverts commit 89cb080b2aec2a27ca117002e05452e86b7317be.
This patch relies on the "lib_ist.yaml" file, but there is an issue with
the Windows loader. When the environment variable "VALIDATE_ARGS" is set to "TRUE"
in Windows, it results in a double string value ('"True"'). In contrast, on Linux,
it is treated as a single string value. This discrepancy is preventing the generation
of the "lib_list.yaml" file on Windows, leading to a failure.

Signed-off-by: Sathish Kumar Kamishettigari <sathishkumar.kamishettigari@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>