dp21txss: Fix YUV format issues in SST mode
When DP21 is set up in SST mode, the dynamic range variable is cleared in sst mode,
which is incorrect, and the video displays a blank screen when Tx attempts to
transmit in 444 and 422 formats.
Thus, the clearing the variable of dynamic range is removed.

Signed-off-by: Lakshmi Prasanna Eachuri <lakshmi.prasanna.eachuri@amd.com>

Acked-for-series: Gaddipati, Naveen <naveen.gaddipati@amd.com>