xilsfl: src: Enable copying of flash ID for QSPI flashes
The current implementation copies flash ID data to the driver
instance's DeviceIdData for OSPI flash, but this leads to tuning
failures in SDR-PHY mode when using QSPI flash, as the driver
instance's DeviceIdData remains empty, leaving nothing to compare
the flash ID against during tuning.

To resolve this issue, remove the condition that restricts
copying the flash ID to the driver instance, irrespective of
the flash type.

Signed-off-by: Srikanth Boyapally <srikanth.boyapally@amd.com>

Acked-for-series: Srinivas Goud <srinivas.goud@amd.com>