SpartanUP: xilpuf: Align printed PUF Helper Data size to 516 bytes
In SpartanUp, the space reserved for PUF helper data in bootheader is
516 bytes i.e 129 words (127 words of syndrome data, 1 word for CHASH
and 1 word for Aux). Currently the size of the helper data printed by
the example is 528 bytes i.e. 132 words as it includes the padding bytes
required to calculate the SHA3 hash. This is not user friendly as the user
will have to remove the padding bytes before placing it in the bootheader.

Hence updated the example to print 516 bytes of PUF helper data.

Signed-off-by: Mounika Botcha <mounika.botcha@amd.com>
Acked-by: Harsha <harsha.harsha@amd.com>