xilloader: Reset SHA instance 1 after LoadImage
During the load of PCR enabled secure partial PDI, if it
fails for some reason SHA instance 1 is held in non-reset state and
due to which next consequent PDI loads are getting failed for KAT/SHA3 1
operation.
This is fixed by resetting the SHA instance 1 after LoadImage
irrespective of success/failure of image load.

Signed-off-by: Kalyani Akula <kalyani.akula@amd.com>
Acked-by: Durga Challa <durga.challa@amd.com>