xilpm: versal_common: Change core state only if core is running
Currently, subsystem restart fails after a kernel crash because it sets the
core state to XPM_DEVSTATE_PENDING_PWR_DWN even when it was already powered
down, which prevents proper shutdown.

Now, the core state is updated only if the core is actually running, allowing
the subsystem to restart correctly.

Signed-off-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>
Acked-by: Naman Trivedi <naman.trivedimanojbhai@amd.com>