scripts:pyesw: Regenerate lib_list.yaml for BSP compatibility with older flow
This is a workaround for the issue where the BSP does not work with the older flow.
The patch checks if lib_list.yaml exists and contains the example key. If not,
it regenerates lib_list.yaml to ensure the required structure is present,
This step is necessary for the BSP to function correctly with the older flow.

Signed-off-by: Sathish Kumar Kamishettigari <sathishkumar.kamishettigari@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>