zynqmp_pmufw: fix canonical PMU register defines
The logic for creating canonical PMU register definitions made incorrect
assumptions. For instance, it assumed that XPAR_XZDMA_0_BASEADDR would
always be 0xFFA80000U, which is not guaranteed.

This patch corrects the logic to check against all possible base addresses
for a given peripheral and defines the corresponding canonical PMU register
macros accordingly.

Signed-off-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>
Acked-by: Ronak Jain <ronak.jain@amd.com>