xilpm: versal_common: server: Remove xpm_plat_proc.c and xpm_plat_proc.h file
The XPmPlatAddProcDevice function was introduced during the Telluride
refactoring but is no longer used in the current xilpm codebase. Although
it returned XST_FAILURE, the correct behavior would have been to return
XST_INVALID_PARAM for invalid inputs.

Since the function is now redundant and unreferenced, it has been removed
to simplify the codebase. This cleanup was prompted by unit test coverage
of the add node functionality.

Signed-off-by: Harshil Godhani <HarshilDharmeshbhai.Godhani@amd.com>
Acked-by: Naman Trivedi <naman.trivedimanojbhai@amd.com>