xilocp: Remove depends libraries for XilOcp
Having depends libraries in xilocp yaml is resulting in
compilation issues. So, the guidance is to add all
dependent libraries during the bsp compilation.
Earlier for other secure libraries this is taken care,
missed it for XilOcp. Updated the yaml file to remove
depends libraries.

Signed-off-by: Kalyani Akula <kalyani.akula@amd.com>

Acked-for-series: Durga Challa <durga.challa@amd.com>