xilloader: Secure components extension to PCR for Versal_2Ve_2Vm
SPK hash, SPK ID hash and partition revocation id extension is
added for Versal_2Ve_2Vm.

Due to the complexities of the Telluride PDI format, there are some
limitations on the security information extended to PCR.
1. Single SPK case
HW PCR0 and HW PCR 1 will have the SPK and SPK_ID extended and
the same will be copied to SW PCR by PLM upon boot when Guard bit is
programmed.
2. Only Encryption case
PLM extends Partition RevokeId from the corresponding partition
header only once per Image
3. Single SPK case + Encryption
AC components are extended to SW PCR 0/1 as per #1 and
then Partition RevokeId to corresponding PCR as requested in BIF
4. Multiple SPK case
Recommendation is to use single SPK for a complete subsystem.
When PCR information is enabled for a given subsystem in bif,
PLM extends SPK hash and SPK ID hash once the SPK is authenticated
and after the first chunk is successfully processed.
5. Multiple SPK case + Encryption
When PCR information is enabled for a given subsystem in bif,
PLM extends SPK hash, SPK ID hash and Partition RevokeId once
the SPK is authenticated and after the first chunk is successfully
authenticated and decrypted.

Signed-off-by: Kalyani Akula <kalyani.akula@amd.com>

Acked-for-series: Durga Challa <durga.challa@amd.com>