sw_apps: asufw: Add support for IPI message filtering
Command permissions for all the IPI commands supported by ASU will come
from tools. In this patch, support for the below is added:
1. Add support for giving access permissions per channel basis to the
exisiting commands and assign default permissions
2. During boot, PLM will copy the command permissions to the ASU RTCA
memory. ASUFW reads this data and updates all the modules with the given
access permissions
3. When a command is received, ASUFW will validate the command based on
the access permissions given for the channel with the security flag in
the command header given by the user.
If the command access permissions for the channel is specified as NO
ACCESS or if the security flag and the command access permissions
doesn't match, ASUFW will return an error to the user.

Signed-off-by: Mounika Akula <mounika.akula@amd.com>

Acked-for-series: Durga Challa <durga.challa@amd.com>