sw_services: xilasu: Update client library to include secure flag in command header
Bit 24 in command header is a secure flag which indicates whether the
command is secure or non-secure. Based on this information, ASUFW
processes the request if the access permissions of the channel match
with the secure flag bit. XilAsu and the client examples are updated to
include the secure flag in command header.

Signed-off-by: Mounika Akula <mounika.akula@amd.com>

Acked-for-series: Durga Challa <durga.challa@amd.com>