xilpm: versal_common: server/client: Deprecate XPm_RequestSuspend API
Deprecate the use of the XPm_RequestSuspend API for the Versal and
Versal-Net platforms. This is because the API is intended for
suspending cross-subsystems, and the same functionality can now be
achieved using the XPm_ForcePowerdown API. Therefore, continuing to
use XPm_RequestSuspend API may no longer be necessary. Hence
deprecating the same.

Signed-off-by: Ronak Jain <ronak.jain@amd.com>

Acked-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>