sw_apps: asufw: Fix GCC warnings by explicitly indicating the intentional fall-through
The queue scheduler logic contains stages for each command execution.
This was recently replaced with switch from if-else logic to simplify.
In the new logic, there are intentional fall-through cases to go to the
next stage execution of the command if condition satifies. This is
generating the GCC warning. Hence, fixing the same by specifying
explicitly to indicate the intentional fall-through to the compiler by
adding a comment.

Signed-off-by: Mounika Akula <mounika.akula@amd.com>
Acked-by: Sai Srinivas Sirimisetty <SaiSrinivas.Sirimisetti@amd.com>