xilpm: server: common: Enhance subsystem activation handling
Subsystems are activated before loading starts.  This is important to
ensure that memories and processors are ready for the load to start.
However, if the subsystem depends on boot devices to be pre-allocated,
then it is not possible to activate the subsystem since the boot device
is still required by firmware to do the actual loaded.  This is
independent of hand-off order.

Change PM subsystem activation such that processors and memories are
activated before subsystem load starts, and the remaining devices are
activated immediately before hand-off. This is achieved by enhancing
subsystem activation function to take an action -
PREALLOC_PROC_MEM_ONLY, PREALLOC_PERIPH_ONLY or PREALLOC_ALL. Subsystem
is considered activated only after all processors, memories and
peripherals within the subsystem are pre-allocated.

Signed-off-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>
Acked-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>