sw_apps: zynqmp_fsbl: Fix intermittent failures in the POS use case
On ZynqMP platforms, the Power-Off Suspend (POS) feature was observed
to fail intermittently (~50% of the time). The failure was traced to
the system not reaching the WFE (Wait For Event) state in FSBL, which
is required for a successful resume. Instead, FSBL execution was getting
stuck in the XUartPs_WaitTransmitDone() function, thereby occasionally
causing a hang when PMUFW executes the restore/resume process. To fix
this issue for POS use case modified the boot flow to put FSBL in WFE
state without checking UART transmission to done which is not required
for POS use case

Signed-off-by: Dasari Sharath Kumar <sharath.kumar.dasari@amd.com>
Acked-by: Mounika Akula <mounika.akula@amd.com>