xilocp: Correct the value of the XOCP_EFUSE_PPK_NUM_OF_BYTES
Only the upper 256 bits of the SHA3-384 hash of the PPK are
written to eFuse.SWPCR1 is extended with the PPK hash from eFuses.
However, while copying the PPK hash, a size of 48 bytes was used
instead of 32 bytes, leading to a potential overflow.
The macro is updated to reflect the correct size of the eFuse PPK hash.

Signed-off-by: Kalyani Akula <kalyani.akula@amd.com>

Acked-for-series: Durga Challa <durga.challa@amd.com>