xilloader: Point the correct hashblock for getting partition hash
The measurement logic for versal_2ve_2vm currently retrieves the
partition hash from the meta header hash block.
However, for authenticated PDIs with multiple SPKs, all partition
hashes are stored in the HashBlock instance, not in the meta header
hash block. Update the code to reference the correct HashBlock
instance for partition hashes.

Signed-off-by: Kalyani Akula <kalyani.akula@amd.com>
Acked-by: Durga Challa <durga.challa@amd.com>