xilloader: Initialize DigestIndex before partitions load
For authenticated or enccrypted PDIs, when PCR is enabled
measurement index is increamented internally to extend
secure components in sequence like for auth pdi SPK key hash,
SPK ID hash and then Image Hash.
For this to happen measurment index has to be initialilzed
before partition loading starts and send the reference of the
measurement index to secure config measurement API where it
will extend the secure components and increament the index
and store it in PDI instance.

Signed-off-by: Kalyani Akula <kalyani.akula@amd.com>
Acked-by: Durga Challa <durga.challa@amd.com>