xilpm_ng: Fix subsystem max index tracking
Subsystem max index was incorrectly tracked using the number of
subsystems. Fix this by adding a dedicated field to track the maximum
index of the subsystems added. This ensures correct behavior when there
are subsystems with non-sequential IDs.

Also, update the condition in XPm_HookAfterPlmCdo to check for when to
add the requirements of the default subsystem based on the new max
index. When default subsystem is the only one present, its index should
be the max (=0), as PMC and ASU subsystems are not added as subsystems,
yet.

Signed-off-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>

Acked-for-series:Trung Tran<trung.tran@amd.com>