xilpm_ng: Enable support for boot/recovery timer for subsystem health
Enable healthy boot timer support for:
 1. Boot time subsystem health/load monitoring
 2. Runtime recovery monitoring during subsystem restart/force powerdown

This helps in monitoring the Linux hang cases and/or TF-A hang cases and
enables an escalated path for recovery of the system.

In case of linux subsystem restart, this feature is useful as it can
help complete the graceful powerdown even when linux is not responding
but TF-A is still alive.

To support this cleanly, introduce a new FSM (as opposed to using a
generic device FSM - as the functionality here is specific to healthy
boot monitor devices). The new FSM will handle the healthy boot monitor
request/release operations.

This functionality is only supported in SUBSYS module and not EEMI
module, as healthy boot monitoring can only be enabled when there are
user defined subsystems in the solution.

Remove generic timer ops from the subsystem ops as they are not
required.

Signed-off-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>

Acked-for-series:Trung Tran<trung.tran@amd.com>