xilpm_ng: subsys: Fix setting allowed subsystems for reset nodes
When setting allowed subsystems for a reset node, the existing
implementation overrides the previous mask with the new one.
If the reset is shared among multiple subsystems, this will only
allow the last caller's subsystems to be set, effectively removing
previously set permissions.

Fix this by updating the existing allowed subsystems mask through a
bitwise OR operation.

Signed-off-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@amd.com>

Acked-by: Kush Jain <kush.jain@amd.com>