clk_wiz: Conditionally compile XClk_Wiz_GetInputRate
The input rate detection is supported only in clkx5. In the
xsct flow there is no way to differenciate the clkx5. For SDT
we are using the compatible to differenciate. So this feature
are not supported in xsct flow.

Compile XClk_Wiz_GetInputRate conditionally. While at it also
change the Assert to nonvoid.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
Acked-by: Srinivas Goud <srinivas.goud@amd.com>