dcsub: examples: Fix header file references for 4K HBR3 and SDP examples
The 4K HBR3 and SDP examples fail to build due to incorrect header file
references in their respective source files. This issue was introduced
when adding the missing examples to the Vitis GUI but the header file
names were not updated accordingly.

This patch updates the affected source files to use the correct header
file names, ensuring the examples build successfully.

Signed-off-by: Akhil Raghu Menon <akhil.raghu.menon@amd.com>

Acked-by: Vishal Sagar <vishal.sagar@amd.com>