sw_services:xilloader: Fix non-secure boot failure on crypto disabled boards when PLM_OCP_KEY_MNGMT is disabled
The current implementation performs data measurement even when the user has not requested PCR extension.
As SHA is used for data measurement and is unavailable on crypto-disabled boards, boot is getting halted.
This patch handles this case when PLM_OCP_KEY_MNGMT is OFF by checking the PCR info value before proceeding to the data measurement.
The fix is done for VersalNet, Versal_2VP and Versal_2VP_P platforms.

Signed-off-by: Preethi Kolli <preethi.kolli@amd.com>
Acked-by: Kalyani Akula <kalyani.akula@amd.com>